+++ ntsc-mtl-demo/Filters/TextureFilters/IIRTextureFilter.swift
@@ -21,7 +21,7 @@ class IIRTextureFilter {
     enum InitialCondition {
         case zero
         case firstSample
-        case constant([Float])
+        case constant([Float16])
     }
     
     private static var iirInitialConditionPipelineState: MTLComputePipelineState?
@@ -33,21 +33,21 @@ class IIRTextureFilter {
     
     private let device: MTLDevice
     private let library: MTLLibrary
-    private let numerators: [Float]
-    private let denominators: [Float]
+    private let numerators: [Float16]
+    private let denominators: [Float16]
     private let initialCondition: InitialCondition
     private let channelMix: YIQChannels
-    private let scale: Float
+    private let scale: Float16
     private(set) var zTextures: [MTLTexture] = []
     private var scratchTexture: MTLTexture?
     
-    init(device: MTLDevice, library: MTLLibrary, numerators: [Float], denominators: [Float], initialCondition: InitialCondition, channels: YIQChannels, scale: Float, delay: UInt) {
+    init(device: MTLDevice, library: MTLLibrary, numerators: [Float16], denominators: [Float16], initialCondition: InitialCondition, channels: YIQChannels, scale: Float16, delay: UInt) {
         self.device = device
         self.library = library
         let maxLength = max(numerators.count, denominators.count)
-        var paddedNumerators: [Float] = Array(repeating: 0, count: maxLength)
+        var paddedNumerators: [Float16] = Array(repeating: 0, count: maxLength)
         paddedNumerators[0..<numerators.count] = numerators[0...]
-        var paddedDenominators: [Float] = Array(repeating: 0, count: maxLength)
+        var paddedDenominators: [Float16] = Array(repeating: 0, count: maxLength)
         paddedDenominators[0..<denominators.count] = denominators[0...]
         self.numerators = paddedNumerators
         self.denominators = paddedDenominators
@@ -80,17 +80,17 @@ class IIRTextureFilter {
         initialCondition: InitialCondition,
         initialConditionTexture: MTLTexture,
         textures: [MTLTexture],
-        numerators: [Float],
-        denominators: [Float],
+        numerators: [Float16],
+        denominators: [Float16],
         library: MTLLibrary,
         device: MTLDevice,
         commandBuffer: MTLCommandBuffer
     ) throws {
         let region = MTLRegionMake2D(0, 0, outputTexture.width, outputTexture.height)
-        let bytesPerRow: Int = MemoryLayout<Float>.size * 4 * outputTexture.width
+        let bytesPerRow: Int = MemoryLayout<Float16>.size * 4 * outputTexture.width
         switch initialCondition {
         case .zero:
-            let input: [Float] = [0, 0, 0, 1]   // black/zero
+            let input: [Float16] = [0, 0, 0, 1]   // black/zero
             var yiqa = input
             for tex in textures {
                 tex.replace(region: region, mipmapLevel: 0, withBytes: &yiqa, bytesPerRow: bytesPerRow)
@@ -118,18 +118,18 @@ class IIRTextureFilter {
             den / firstNonZeroCoeff
         }
         
-        var bSum: Float = 0
+        var bSum: Float16 = 0
         for i in 1 ..< numerators.count {
             let num = normalizedNumerators[i]
             let den = normalizedDenominators[i]
             bSum += num - (den * normalizedNumerators[0])
         }
         let z0Fill = bSum / normalizedDenominators.reduce(0, +)
-        let z0FillValues: [Float] = [z0Fill, z0Fill, z0Fill, 1]
+        let z0FillValues: [Float16] = [z0Fill, z0Fill, z0Fill, 1]
         var z0FillMutable = z0FillValues
         textures[0].replace(region: region, mipmapLevel: 0, withBytes: &z0FillMutable, bytesPerRow: bytesPerRow)
-        var aSum: Float = 1
-        var cSum: Float = 0
+        var aSum: Float16 = 1
+        var cSum: Float16 = 0
         for i in 1 ..< numerators.count {
             let num = normalizedNumerators[i]
             let den = normalizedDenominators[i]
@@ -154,7 +154,7 @@ class IIRTextureFilter {
         )
     }
     
-    private static func initialConditionFill(textureToFill: MTLTexture, initialConditionTexture: MTLTexture, aSum: Float, cSum: Float, library: MTLLibrary, device: MTLDevice, commandBuffer: MTLCommandBuffer) throws {
+    private static func initialConditionFill(textureToFill: MTLTexture, initialConditionTexture: MTLTexture, aSum: Float16, cSum: Float16, library: MTLLibrary, device: MTLDevice, commandBuffer: MTLCommandBuffer) throws {
         let pipelineState: MTLComputePipelineState
         if let iirInitialConditionPipelineState {
             pipelineState = iirInitialConditionPipelineState
@@ -174,9 +174,9 @@ class IIRTextureFilter {
         commandEncoder.setTexture(textureToFill, index: 0)
         commandEncoder.setTexture(initialConditionTexture, index: 1)
         var aSum = aSum
-        commandEncoder.setBytes(&aSum, length: MemoryLayout<Float>.size, index: 0)
+        commandEncoder.setBytes(&aSum, length: MemoryLayout<Float16>.size, index: 0)
         var cSum = cSum
-        commandEncoder.setBytes(&cSum, length: MemoryLayout<Float>.size, index: 1)
+        commandEncoder.setBytes(&cSum, length: MemoryLayout<Float16>.size, index: 1)
         commandEncoder.dispatchThreads(
             MTLSize(width: textureToFill.width, height: textureToFill.height, depth: 1),
             threadsPerThreadgroup: MTLSize(width: 8, height: 8, depth: 1))
@@ -304,7 +304,7 @@ class IIRTextureFilter {
     static func finalImage(
         inputImage: MTLTexture,
         filteredImage: MTLTexture,
-        scale: Float,
+        scale: Float16,
         library: MTLLibrary,
         device: MTLDevice,
         commandBuffer: MTLCommandBuffer
@@ -327,7 +327,7 @@ class IIRTextureFilter {
         commandEncoder.setTexture(inputImage, index: 0)
         commandEncoder.setTexture(filteredImage, index: 1)
         var scale = scale
-        commandEncoder.setBytes(&scale, length: MemoryLayout<Float>.size, index: 0)
+        commandEncoder.setBytes(&scale, length: MemoryLayout<Float16>.size, index: 0)
         commandEncoder.dispatchThreads(
             MTLSize(width: inputImage.width, height: inputImage.height, depth: 1),
             threadsPerThreadgroup: MTLSize(width: 8, height: 8, depth: 1))
@@ -361,7 +361,7 @@ class IIRTextureFilter {
         
         
         var channelMix = channels.floatMix
-        commandEncoder.setBytes(&channelMix, length: MemoryLayout<Float>.size * 4, index: 0)
+        commandEncoder.setBytes(&channelMix, length: MemoryLayout<Float16>.size * 4, index: 0)
         commandEncoder.dispatchThreads(
             MTLSize(width: inputImage.width, height: inputImage.height, depth: 1),
             threadsPerThreadgroup: MTLSize(width: 8, height: 8, depth: 1))
@@ -370,7 +370,7 @@ class IIRTextureFilter {
     
 
     
-    static func sideEffect(inputImage: MTLTexture, z: MTLTexture, zPlusOne: MTLTexture, filteredSample: MTLTexture, numerator: Float, denominator: Float, library: MTLLibrary, device: MTLDevice, commandBuffer: MTLCommandBuffer) throws {
+    static func sideEffect(inputImage: MTLTexture, z: MTLTexture, zPlusOne: MTLTexture, filteredSample: MTLTexture, numerator: Float16, denominator: Float16, library: MTLLibrary, device: MTLDevice, commandBuffer: MTLCommandBuffer) throws {
         let pipelineState: MTLComputePipelineState
         if let iirSideEffectPipelineState {
             pipelineState = iirSideEffectPipelineState
@@ -392,15 +392,15 @@ class IIRTextureFilter {
         commandEncoder.setTexture(filteredSample, index: 3)
         var num = numerator
         var denom = denominator
-        commandEncoder.setBytes(&num, length: MemoryLayout<Float>.size, index: 0)
-        commandEncoder.setBytes(&denom, length: MemoryLayout<Float>.size, index: 1)
+        commandEncoder.setBytes(&num, length: MemoryLayout<Float16>.size, index: 0)
+        commandEncoder.setBytes(&denom, length: MemoryLayout<Float16>.size, index: 1)
         commandEncoder.dispatchThreads(
             MTLSize(width: inputImage.width, height: inputImage.height, depth: 1),
             threadsPerThreadgroup: MTLSize(width: 8, height: 8, depth: 1))
         commandEncoder.endEncoding()
     }
         
-    static func filterSample(_ inputTexture: MTLTexture, zTex0: MTLTexture, filteredSampleTexture: MTLTexture, num0: Float, library: MTLLibrary, device: MTLDevice, commandBuffer: MTLCommandBuffer) throws {
+    static func filterSample(_ inputTexture: MTLTexture, zTex0: MTLTexture, filteredSampleTexture: MTLTexture, num0: Float16, library: MTLLibrary, device: MTLDevice, commandBuffer: MTLCommandBuffer) throws {
         let pipelineState: MTLComputePipelineState
         if let iirFilterSamplePipelineState {
             pipelineState = iirFilterSamplePipelineState
@@ -421,7 +421,7 @@ class IIRTextureFilter {
         encoder.setTexture(zTex0, index: 1)
         encoder.setTexture(filteredSampleTexture, index: 2)
         var num0 = num0
-        encoder.setBytes(&num0, length: MemoryLayout<Float>.size, index: 0)
+        encoder.setBytes(&num0, length: MemoryLayout<Float16>.size, index: 0)
         encoder.dispatchThreads(
             MTLSize(width: inputTexture.width, height: inputTexture.height, depth: 1),
             threadsPerThreadgroup: MTLSize(width: 8, height: 8, depth: 1))
